# Leo's populate_hits_avx