@lemire 2019-09-30
cc @searchivarius